requirements.txt
Django==4.0.5
python-telegram-bot==13.12
python_dotenv==0.20.0
requests==2.28.1
beautifulsoup4==4.11.1
python-telegram-bot-pagination==0.0.2
loguru==0.6.0
pathvalidate==2.5.2
urllib3==1.26.12
django-import-export==2.9.0